[PreProcessing] Preprocessing before running the image to ML

+++ OCR/Kernels.h
@@ -1,10 +1,8 @@
 #ifndef KERNELS_H
 #define KERNELS_H
 
-#include <stdlib.h>
-
-extern float gauss[9];
-extern float sharpen[9];
-extern float edge_detection[9];
+extern double gauss[9];
+extern double sharpen[9];
+extern double edge_detection[9];
 
 #endif
